refactor: Refactor account routes and data classes
- Changed route path from "/account" to "/accounts" for consistency.
- Refactored data classes into separate file for better organization.
- Updated error message in response to fix typo.
