+++ src/main/kotlin/gay/extremist/models/Account.kt
@@ -35,15 +35,4 @@ class Account(id: EntityID<Int>): Entity<Int>(id) {
     var followedTags by Tag via AccountFollowsTag
 }
 
-@Serializable
-data class LoginAccount(val username: String, val password: String)
-
-@Serializable
-data class RegisterAccount(val username: String, val email: String, val password: String)
-
-@Serializable
-data class RegisteredAccount(val accountId: Int, val token: String)
-
-@Serializable
-data class UnprivilegedAccount(val accountID: Int, val username: String, val email: String)
